Update README.md
This expands the readme.md to serve as a resource giving an overview of all the related components of the ongoing php related initiatives tied to "servehappy".  Intended to be a single resource that can be linked to (already linked to via the #core-php slack channel topic) for helping people get up to speed on what the "servehappy" project is all about and where to find all related initiatives.